Upgrade speed effect might be a fraction.

+++ src/main/java/org/joelson/cts/calculator/model/GeneratorState.java
@@ -1,6 +1,6 @@
 package org.joelson.cts.calculator.model;
 
-public record GeneratorState(int count, float efficiency, int speed) {
+public record GeneratorState(int count, float efficiency, float speed) {
 
     public static GeneratorState EMPTY = new GeneratorState(0, 1);
     public static GeneratorState EMPTY_TIMED = new GeneratorState(0, 1, 1);
