+++ src/main/java/org/joelson/cts/calculator/model/UpgradeEffect.java
@@ -1,6 +1,6 @@
 package org.joelson.cts.calculator.model;
 
-public record UpgradeEffect(Generator generator, float efficiency, int speed) {
+public record UpgradeEffect(Generator generator, float efficiency, float speed) {
 
     public UpgradeEffect(Generator generator, float efficiency) {
         this(generator, efficiency, 1);
