+++ src/main/java/org/joelson/cts/calculator/model/builder/UpgradeBuilder.java
@@ -24,7 +24,7 @@ public record UpgradeBuilder(GardenBuilder gardenBuilder, Generator generator, U
         return gardenBuilder.addGeneratorUpgrade(generator, name, cost, efficiency);
     }
 
-    public UpgradeBuilder addUpgrade(String name, Amount cost, float efficiency, int speed) {
+    public UpgradeBuilder addUpgrade(String name, Amount cost, float efficiency, float speed) {
         return gardenBuilder.addGeneratorUpgrade(generator, name, cost, efficiency, speed);
     }
 }
